add a navigation container and home page navigator
